libifupdown: check executor exit status in handle_executors_for_phase
Without this patch, handle_executors_for_phase would always return true,
even if the executor exited with a non-zero exit status. I noticed this
in conjunction with the dhcp executor were libifupdown would incorrectly
assume that the executor was executed successfully if no lease was
acquired. This caused ifupdown to get into a weird state where it
assumed that the interface was configured successfully even though it
wasn't, thus requiring an explicit `ifdown <interface>` before
attempting to acquire a lease again with `ifup <interface>`.