Add support for optional netlink state detection via libmnl
This commits adds a new waitif API to libifupdown which allows
monitoring the interface state via the Linux netlink subsystem. The
interface monitor is implemented as a background thread which monitors
events received from a netlink socket using libmnl.

When running executors, this waitif API can be used to block after the
pre-up phase until the interface, that is currently being configured,
actually changed to the desired IFF_UP / IFF_RUNNING state. Thereby,
this changes made in this commit ensure that executors of the up phase
are only run after the interface was up'ed successfully.

This new API is optional and deactivated by default.

Alternatives: This feature could have also been implemented as a
external executor, however, an implementation within libifupdown
is easier and more efficient.

Fixes #178