ifupdown: continue bringing up interfaces if state change failed
Without this patch, `ifup -a` would stop bringing up interfaces
if a single interface state change failed. Combined with the
previous commit this may cause `ifup -a` to leave the machine
in an offline state if a single up executor failed for any given
interface. With this commit, `ifup -a` will continue attempting
to bring up other interfaces.