Load the dummy kernel module
A dummy0 interface is automatically created when loading the dummy kernel module. This
presents a problem should the kernel module be loaded as a result of a
`ip link add dummy0 type dummy`, as the command fails due to dummy0 suddenly existing
as a result of the integrated kernel module load.

By loading the kernel module separately, the dummy0 interface is detected and reused
should dummy0 be specified as the interface name with ifupdown-ng.