networking.initd: Cleanup `stop`
This too looses the redirection to `/dev/null`.
In my opinion, nothing in `ifupdown` or its executors should print to `stdout` unless it's important.
Currently the only thing that does, that I know of, is `udhcpc` with a single "OK" once it's done. That should be redirected away at the executor level, not by a service script.